+++ apps/formio-server/src/middleware/upload.js
@@ -1,11 +1,11 @@
 'use strict';
 const fetch = require('formio/src/util/fetch');
 const {promisify} = require('util');
-const PDF_SERVER = process.env.PDF_SERVER || process.env.FORMIO_FILES_SERVER || 'https://files.form.io';
 const _ = require('lodash');
 const Promise = require('bluebird');
 const fs = require('fs');
 const {getLicenseKey} = require('../util/utilization');
+const {getPDFUrls} = require('../util/pdf');
 const debug = require('debug')('formio:pdf:upload');
 const FormData = require('form-data');
 
@@ -27,15 +27,11 @@ module.exports = (formioServer) => async (req, res, next) => {
 
   try {
     // Load project
-    const project = req.primaryProject;
+    const project = req.currentProject;
 
     // Set the files server
-    let filesServer = PDF_SERVER;
-    if (process.env.FORMIO_HOSTED && project.settings.pdfserver) {
-      // Allow them to download from any server if it is set to the default
-      filesServer = project.settings.pdfserver;
-    }
-    debug(`FileServer: ${filesServer}`);
+    const pdfUrls = getPDFUrls(project);
+    debug(`FileServer: ${pdfUrls.local}`);
 
     // Create the headers object
     const headers = {'x-license-key': getLicenseKey(req)};
@@ -63,7 +59,7 @@ module.exports = (formioServer) => async (req, res, next) => {
     }
 
     try {
-      debug('POST: ' + `${filesServer}/pdf/${pdfProject}/file`);
+      debug('POST: ' + `${pdfUrls.local}/pdf/${pdfProject}/file`);
       debug(`Filepath: ${  req.files.file.path}`);
       const form = new FormData();
       form.append('file', fs.createReadStream(req.files.file.path), {
@@ -71,7 +67,7 @@ module.exports = (formioServer) => async (req, res, next) => {
         contentType: req.files.file.type,
         size: req.files.file.size,
       });
-      fetch(`${filesServer}/pdf/${pdfProject}/file`, {
+      fetch(`${pdfUrls.local}/pdf/${pdfProject}/file`, {
         method: 'POST',
         headers: headers,
         body: form,
@@ -83,17 +79,15 @@ module.exports = (formioServer) => async (req, res, next) => {
           res.status(400).send(err.message || err);
           throw err;
         })
-        .then((response) => {
+        .then(async (response) => {
           if (response.ok) {
-            return response.json();
+            const body = await response.json();
+            body.filesServer = pdfUrls.public;
+            return res.status(201).send(body);
           }
-          if (response.statusCode > 299) {
-            res.status(response.statusCode).send(response.body);
+          else {
+            return res.status(response.status).send(await response.text());
           }
-        })
-        .then(async (body) => {
-          body.filesServer = filesServer;
-          res.status(201).send(body);
         });
     }
     catch (err) {