+++ apps/formio-server/src/plans/Trial.js
@@ -10,6 +10,7 @@ module.exports = class Trial extends Plan {
       submissionRequests: 10000,
       emails: 100,
       stages: 5,
+      livestages: 2,
       plan: 'trial',
     };
   }