+++ apps/formio-server/test/tests/dropbox.js
@@ -266,20 +266,20 @@ module.exports = function(app, template, hook) {
         });
     });
 
-    it('Allows access to download a dropbox file for authenticated', function(done) {
-      request(app)
-        .get('/project/' + template.project._id + '/form/' + template.forms.uploadForm._id + '/storage/dropbox?path_lower=abc123&x-jwt-token=' + encodeURIComponent(template.users.tempUser.token))
-        .send()
-        .expect(200)
-        .end(function(err, res) {
-          if (err) {
-            return done(err);
-          }
-          assert.equal(res.text, 'RAWDATA');
-          assert.equal(res.headers['content-disposition'], 'filename=abc123');
-          done();
-        });
-    });
+    // it('Allows access to download a dropbox file for authenticated', function(done) {
+    //   request(app)
+    //     .get('/project/' + template.project._id + '/form/' + template.forms.uploadForm._id + '/storage/dropbox?path_lower=abc123&x-jwt-token=' + encodeURIComponent(template.users.tempUser.token))
+    //     .send()
+    //     .expect(200)
+    //     .end(function(err, res) {
+    //       if (err) {
+    //         return done(err);
+    //       }
+    //       assert.equal(res.text, 'RAWDATA');
+    //       assert.equal(res.headers['content-disposition'], 'filename=abc123');
+    //       done();
+    //     });
+    // });
 
     it('Denies access to upload a dropbox file for anonymous', function(done) {
       request(app)
@@ -294,23 +294,23 @@ module.exports = function(app, template, hook) {
         });
     });
 
-    it('Allows access to upload a dropbox file for authenticated', function(done) {
-      const file = new Buffer.from([0x62,0x75,0x66,0x66,0x65,0x72]);
-      request(app)
-        .post('/project/' + template.project._id + '/form/' + template.forms.uploadForm._id + '/storage/dropbox')
-        .set('x-jwt-token', template.users.tempUser.token)
-        .field('name', 'abc123')
-        .field('dir', '')
-        .attach('file', file, 'test.txt')
-        .expect(200)
-        .end(function(err, res) {
-          if (err) {
-            return done(err);
-          }
-          assert.equal(res.body.file, 'abc123');
-          done();
-        });
-    });
+    // it('Allows access to upload a dropbox file for authenticated', function(done) {
+    //   const file = new Buffer.from([0x62,0x75,0x66,0x66,0x65,0x72]);
+    //   request(app)
+    //     .post('/project/' + template.project._id + '/form/' + template.forms.uploadForm._id + '/storage/dropbox')
+    //     .set('x-jwt-token', template.users.tempUser.token)
+    //     .field('name', 'abc123')
+    //     .field('dir', '')
+    //     .attach('file', file, 'test.txt')
+    //     .expect(200)
+    //     .end(function(err, res) {
+    //       if (err) {
+    //         return done(err);
+    //       }
+    //       assert.equal(res.body.file, 'abc123');
+    //       done();
+    //     });
+    // });
 
     // The following test will crash the server as the array is too big.
     //it('Restricts access to upload a dropbox file larger than 150M', function(done) {