+++ apps/formio-server/Changelog.md
@@ -8,16 +8,127 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - User Sessions
  - Isomorphic Validation
- - New Licensing system.
+ - Audit logging
+ - New Licensing system
+ - Stage Types (Authoring and Live)
+ - Group Roles and improvements to Group Assignment handling.
  
 ### Changed
  - Removed request.js and restler and replaced with node-fetch
  
 ### Removed
+ - Method Override header functionality
  - Jira Action
  - Hubspot Action
  - Office365 Actions
  - Facebook, LinkedIn, Twitter, Dropbox OAuth providers
+
+## 7.0.0-rc.32
+### Changed
+ - Upgrade portal@7.1.17-rc.9
+ - Upgrade formio.js@4.12.0-rc.16
+
+## 7.0.0-rc.31
+### Changed
+ - Upgrade formio@2.0.0-rc.14
+ - Upgrade portal@7.1.17-rc.7
+
+## 7.0.0-rc.28
+### Fixed
+ - Issue where portal was not accessible due to Helmet upgrade.
+
+## 7.0.0-rc.27
+### Changed
+ - Upgrade portal@7.1.17-rc.4
+
+## 7.0.0-rc.26
+### Changed
+ - Upgrade formio@2.0.0-rc.13
+
+## 7.0.0-rc.25
+### Fixed
+ - Attach as PDF to work inline without the need for Save Submission. 
+
+## 7.0.0-rc.19
+### Added
+ - Feat: allowed admins pass "noValidate" flag when they create submissions
+
+### Fixed
+ - Fixes an issue related to externalIds in Webhooks.
+ - FOR-2727: Change the message to upgrade
+ - FOR-2699: added storageAccessHanler in init script
+ 
+### Changed
+ - Upgrade formio@2.0.0-rc.10
+ - Upgrade formiojs@4.12.0-rc.7
+
+## 6.10.8-rc.4
+### Changed
+ - Upgrade portal@7.0.38-rc.6
+ - Upgrade formiojs@4.11.1-rc.6
+
+## 6.10.8-rc.3
+### Changed
+ - Upgrade portal@7.0.38-rc.5
+
+## 6.10.8-rc.2
+### Changed
+ - Upgrade portal@7.0.38-rc.4
+ - Upgrade formiojs@4.11.1-rc.5
+
+### Fixed
+ - Passing project to pdf server to solve image displays.
+
+## 6.10.8-rc.1
+## 6.10.7-rc.2
+### Changed 
+ - Upgrade portal@7.0.38-rc.2
+ - Upgrade formiojs@4.11.1-rc.3
+
+## 6.10.7-rc.1
+### Changed 
+ - Upgrade portal@7.0.38-rc.1
+ - Upgrade formiojs@4.11.1-rc.1
+
+## 6.10.7
+### Changed
+ - Upgrade formio to 1.90.5 to fix mongo ssl connections.
+
+## 6.10.6
+### Changed 
+ - Upgrade formio to 1.90.2 to fix email crash on large emails.
+
+## 6.10.5
+### Changed 
+ - Upgrade formio to 1.90.1 to upgrade resourcejs to 2.3.1 to fix issue with limit and sort.
+
+## 6.10.4
+### No changes: Official build
+
+## 6.10.4-rc.2
+### Changed
+ - Upgrade portal@7.0.36-rc.2
+ - Upgrade formiojs@4.11.0-rc.4
+
+## 6.10.4-rc.1
+### Changed
+ - Upgrade dependencies
+ - Upgrade portal@7.0.36-rc.1
+ - Upgrade formiojs@4.11.0-rc.2
+ - Upgrade formio@1.90.0
+
+### Fixed
+ - FOR-2682: Fixes an issue where x-jwt-token from SAML was too big
+ - fix: added next function calls
+
+## 6.10.3
+### Changed
+ - Upgrade portal@7.0.35
+
+## 6.10.2
+### Changed
+ - Upgrade formiojs@4.10.5
+ - Upgrade portal@7.0.34
 
 ## 6.10.2-rc.7
 ### Changed